requirements.txt
cmake
easydict~=1.9
numpy==1.23
numba
opencv-python~=4.5.1.48
matplotlib~=3.3.3
scipy
scikit-image
tqdm
PyYAML
open3d
cython
tensorboardX
shapely==1.7.1
einops
timm
cumm==0.2.9
tikzplotlib
